requirements.txt
bip-utils==2.3.0
Django>=5.0.8
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
dj-database-url==0.5.0
django-celery-results>=2.4.0
django-celery-beat==2.7.0
django-bmemcached==0.3.0
django-cors-headers==3.12.0
django-encrypted-model-fields==0.6.5
django-filter==21.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
drf-yasg==1.20.0
ed25519==1.5
locust==2.26.0
redis==4.6
python-dotenv==0.20.0
python-memcached==1.59
# sentry-sdk==1.5.12
web3>=7.3.0,<8
coverage~=7.3.2
pytz~=2023.3.post1
requests>=2.32.0
celery~=5.4.0
celery-redbeat==2.2.0
# django-safedelete~=1.3.3
pillow==10.4.0
channels[daphne]==4.1
channels_redis==4.2
django-cloudflare-images~=0.6.0
PyJWT==2.9
python-decouple==3.8
psycopg[binary]==3.2.2
whitenoise[gzip]==6.1.0
drf-spectacular==0.27.2

siwe==4.3.0